+++ test/nextjs/templates.test.ts
@@ -50,6 +50,10 @@ describe('Next.js code templates', () => {
 
           // Setting this option to true will print useful information to the console while you're setting up Sentry.
           debug: false,
+
+          // Enable sending user PII (Personally Identifiable Information)
+          // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+          sendDefaultPii: true,
         });
 
         export const onRouterTransitionStart = Sentry.captureRouterTransitionStart;"
@@ -90,6 +94,10 @@ describe('Next.js code templates', () => {
 
           // Setting this option to true will print useful information to the console while you're setting up Sentry.
           debug: false,
+
+          // Enable sending user PII (Personally Identifiable Information)
+          // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+          sendDefaultPii: true,
         });
 
         export const onRouterTransitionStart = Sentry.captureRouterTransitionStart;"
@@ -120,6 +128,10 @@ describe('Next.js code templates', () => {
 
           // Setting this option to true will print useful information to the console while you're setting up Sentry.
           debug: false,
+
+          // Enable sending user PII (Personally Identifiable Information)
+          // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+          sendDefaultPii: true,
         });
 
         export const onRouterTransitionStart = Sentry.captureRouterTransitionStart;"
@@ -161,6 +173,10 @@ describe('Next.js code templates', () => {
 
           // Setting this option to true will print useful information to the console while you're setting up Sentry.
           debug: false,
+
+          // Enable sending user PII (Personally Identifiable Information)
+          // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+          sendDefaultPii: true,
         });
 
         export const onRouterTransitionStart = Sentry.captureRouterTransitionStart;"
@@ -195,6 +211,10 @@ describe('Next.js code templates', () => {
 
             // Setting this option to true will print useful information to the console while you're setting up Sentry.
             debug: false,
+
+            // Enable sending user PII (Personally Identifiable Information)
+            // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+            sendDefaultPii: true,
           });
           "
         `);
@@ -222,6 +242,10 @@ describe('Next.js code templates', () => {
 
             // Setting this option to true will print useful information to the console while you're setting up Sentry.
             debug: false,
+
+            // Enable sending user PII (Personally Identifiable Information)
+            // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+            sendDefaultPii: true,
           });
           "
         `);
@@ -252,6 +276,10 @@ describe('Next.js code templates', () => {
 
             // Setting this option to true will print useful information to the console while you're setting up Sentry.
             debug: false,
+
+            // Enable sending user PII (Personally Identifiable Information)
+            // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+            sendDefaultPii: true,
           });
           "
         `);
@@ -279,6 +307,10 @@ describe('Next.js code templates', () => {
 
             // Setting this option to true will print useful information to the console while you're setting up Sentry.
             debug: false,
+
+            // Enable sending user PII (Personally Identifiable Information)
+            // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+            sendDefaultPii: true,
           });
           "
         `);
@@ -312,6 +344,10 @@ describe('Next.js code templates', () => {
 
             // Setting this option to true will print useful information to the console while you're setting up Sentry.
             debug: false,
+
+            // Enable sending user PII (Personally Identifiable Information)
+            // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+            sendDefaultPii: true,
           });
           "
         `);
@@ -340,6 +376,10 @@ describe('Next.js code templates', () => {
 
             // Setting this option to true will print useful information to the console while you're setting up Sentry.
             debug: false,
+
+            // Enable sending user PII (Personally Identifiable Information)
+            // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+            sendDefaultPii: true,
           });
           "
         `);
@@ -368,6 +408,10 @@ describe('Next.js code templates', () => {
 
             // Setting this option to true will print useful information to the console while you're setting up Sentry.
             debug: false,
+
+            // Enable sending user PII (Personally Identifiable Information)
+            // https://docs.sentry.io/platforms/javascript/guides/nextjs/configuration/options/#sendDefaultPii
+            sendDefaultPii: true,
           });
           "
         `);